Add files via upload
This is the code for enabling face detection in a video. The only drawback this code faces is that it can detect only one face. need a little improvement in that. before running the code, make sure that there is a folder named known_faces and it contains the directory which contains the face of the person to be recognized. name the folder that containd the photos as their name.